Add initial banking system project files
Added Java classes for Banco, Cliente, Tarjeta, and Logica, as well as a main entry point for a simple banking system. Includes project configuration files for NetBeans and a sample tarjetas.json data file. Implements basic functionality for registering banks, clients, and cards, as well as displaying and updating information.